Add files via upload
The main idea of the project is to show which among the K-NN, SVM, and Neural Networks is a better classifier in the machine learning set and which performs the best on which type of data. The objective is to predict whether a patient has diabetes or not using these three Machine Learning models by factoring in independent variables such as Cholesterol, Smoking Habits, Sex, BMI, Income, Age etc.